+++ crates/gqb/src/prelude.rs
@@ -1,3 +1,6 @@
 //! Convenient import of types for gqb.
 
 pub use crate::utils::ToGraphLabel as _;
+pub use crate::{expression::expression_builder, Builder, Variable};
+
+pub(crate) use crate::expression::Expression;